# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Прогноз отказов оборудования коллекторов",
    "description": "REST API сервиса прогноза отказов датчиков коллекторов на 24 часа (ЛЦТ-2026).\n\n**Вход:** нажмите *Authorize* и введите логин и пароль (в демо-режиме — демо-пользователя из README), поля client_id/secret оставьте пустыми.\n\n**Машина времени:** все запросы на чтение принимают `at`. Демо-данные покрывают 01.07.2026–31.08.2026, например `at=2026-08-01T12:00:00`.\n\n**Симуляция потока** (`/api/sim/*`): проигрывание суток по заранее рассчитанным прогнозам с ускорением; пока она идёт, запросы без `at` работают на модельное время. Онлайн-пересчёт моделью не выполняется.\n\nВремя — московское, без часового пояса.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reasons": {
      "get": {
        "tags": [
          "Сервис и справочники"
        ],
        "summary": "Справочник причин для решений и заявок",
        "operationId": "reasons_api_reasons_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "decision_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "dispatch",
                    "false_alarm",
                    "monitor"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Только причины для этого решения и общие",
              "title": "Decision Type"
            },
            "description": "Только причины для этого решения и общие"
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReasonOut"
                  },
                  "title": "Response Reasons Api Reasons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recommendations": {
      "get": {
        "tags": [
          "Сервис и справочники"
        ],
        "summary": "Справочник рекомендаций для заявок",
        "operationId": "recommendations_api_recommendations_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sensor_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Рекомендации для типа датчика и общие",
              "title": "Sensor Type"
            },
            "description": "Рекомендации для типа датчика и общие"
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecommendationOut"
                  },
                  "title": "Response Recommendations Api Recommendations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dictionaries": {
      "get": {
        "tags": [
          "Сервис и справочники"
        ],
        "summary": "Подписи и значения для фильтров интерфейса",
        "operationId": "dictionaries_api_dictionaries_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dictionaries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sim/state": {
      "get": {
        "tags": [
          "Симуляция потока"
        ],
        "summary": "Состояние симуляции: модельное время, скорость, доступные сутки",
        "description": "Симуляция проигрывает выбранные сутки по **заранее рассчитанным** прогнозам: модельные часы идут с ускорением, на каждом новом часовом срезе новые критические прогнозы уходят в SSE. Онлайн-пересчёт моделью не выполняется. Читать может любая роль — нужно для метки «Симуляция» в интерфейсе. Вызов заодно выполняет такт симуляции (на случай, если планировщик выключен).",
        "operationId": "state_api_sim_state_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimStateOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/sim/start": {
      "post": {
        "tags": [
          "Симуляция потока"
        ],
        "summary": "Запустить симуляцию суток",
        "description": "Симуляция проигрывает выбранные сутки по **заранее рассчитанным** прогнозам: модельные часы идут с ускорением, на каждом новом часовом срезе новые критические прогнозы уходят в SSE. Онлайн-пересчёт моделью не выполняется. Пока симуляция идёт, запросы без `at` работают на модельное время, исходы с незакрытым горизонтом скрыты. Одна симуляция за раз (409), сутки — только с прогнозами (422), скорость ×1…×600. Право sim_control: диспетчер ОДС, инженер данных, администратор. Выключено при SIM_ENABLED=false (403).",
        "operationId": "start_api_sim_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimStateOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/sim/stop": {
      "post": {
        "tags": [
          "Симуляция потока"
        ],
        "summary": "Остановить симуляцию",
        "description": "Идемпотентно: если симуляция не идёт, возвращает состояние без изменений. Право sim_control.",
        "operationId": "stop_api_sim_stop_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimStateOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/settings": {
      "get": {
        "tags": [
          "Настройки"
        ],
        "summary": "Настраиваемые параметры (любая роль — чтение)",
        "operationId": "get_settings__api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Настройки"
        ],
        "summary": "Изменить параметры (только администратор)",
        "description": "Границы уровней риска по вероятности отказа (должны возрастать) и лимиты уведомлений. Действуют сразу, перезапуск не нужен: уровень риска во всех ответах, фильтрах и счётчиках пересчитывается от `prob`, кэш агрегатов сбрасывается. Прогнозы в БД не меняются. Изменение пишется в `audit_log` (`settings_update`).",
        "operationId": "put_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Dictionaries": {
        "properties": {
          "risk_levels": {
            "items": {
              "$ref": "#/components/schemas/RiskLevelOut"
            },
            "type": "array",
            "title": "Risk Levels"
          },
          "decision_types": {
            "items": {
              "$ref": "#/components/schemas/CodeLabel"
            },
            "type": "array",
            "title": "Decision Types"
          },
          "work_order_statuses": {
            "items": {
              "$ref": "#/components/schemas/CodeLabel"
            },
            "type": "array",
            "title": "Work Order Statuses"
          },
          "work_order_priorities": {
            "items": {
              "$ref": "#/components/schemas/CodeLabel"
            },
            "type": "array",
            "title": "Work Order Priorities"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/CodeLabel"
            },
            "type": "array",
            "title": "Roles"
          },
          "system_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "System Types"
          },
          "sensor_types": {
            "items": {
              "$ref": "#/components/schemas/SensorTypeOut"
            },
            "type": "array",
            "title": "Sensor Types"
          }
        },
        "type": "object",
        "required": [
          "risk_levels",
          "decision_types",
          "work_order_statuses",
          "work_order_priorities",
          "roles",
          "system_types",
          "sensor_types"
        ],
        "title": "Dictionaries"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Прогноз не найден"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReasonOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "decision_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "dispatch",
                  "false_alarm",
                  "monitor"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Type",
            "description": "Для какого решения; null — для любого"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "ReasonOut"
      },
...
      "RecommendationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "sensor_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "sensor_type",
          "text"
        ],
        "title": "RecommendationOut"
      },
...
      "RiskLevelOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "color"
        ],
        "title": "RiskLevelOut"
      },
...
      "SettingsIn": {
        "properties": {
          "risk_attention": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Risk Attention",
            "description": "«Внимание» — вероятность не ниже",
            "examples": [
              0.2
            ]
          },
          "risk_risk": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Risk Risk",
            "description": "«Риск» — вероятность не ниже",
            "examples": [
              0.5
            ]
          },
          "risk_critical": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Risk Critical",
            "description": "«Критично» — вероятность не ниже",
            "examples": [
              0.8
            ]
          },
          "notify_cooldown_hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Notify Cooldown Hours",
            "description": "Повторное уведомление по датчику — не чаще, часов",
            "examples": [
              24
            ]
          },
          "notify_per_slice": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Notify Per Slice",
            "description": "Лента и колокольчик: отдельных уведомлений на срез",
            "examples": [
              3
            ]
          },
          "notify_sim_per_slice": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Notify Sim Per Slice",
            "description": "Симуляция: всплывающих уведомлений на срез",
            "examples": [
              5
            ]
          }
        },
        "type": "object",
        "required": [
          "risk_attention",
          "risk_risk",
          "risk_critical",
          "notify_cooldown_hours",
          "notify_per_slice",
          "notify_sim_per_slice"
        ],
        "title": "SettingsIn",
        "description": "Настраиваемые параметры. Границы уровней риска — по вероятности отказа, должны возрастать."
      },
...
      "SimStartIn": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day",
            "description": "Сутки для проигрывания; в базе должны быть прогнозы за них",
            "examples": [
              "2026-06-15"
            ]
          },
          "speed": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Speed",
            "description": "Ускорение: ×60 — сутки за 24 минуты",
            "default": 60,
            "examples": [
              60
            ]
          }
        },
        "type": "object",
        "required": [
          "day"
        ],
        "title": "SimStartIn"
      },
      "SimStateOut": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Симуляция разрешена на сервере (SIM_ENABLED)"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day",
            "description": "Проигрываемые (или последние проигранные) сутки"
          },
          "speed": {
            "type": "integer",
            "title": "Speed"
          },
          "model_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Time",
            "description": "Модельное время: при активной симуляции — «сейчас» для запросов без at"
          },
          "snapshot_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot At",
            "description": "Последний срез прогнозов, по которому разосланы уведомления"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "started_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By"
          },
          "stopped_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason",
            "description": "manual — остановлена вручную, finished — сутки закончились, restart — перезапуск API"
          },
          "events_received": {
            "type": "integer",
            "title": "Events Received",
            "description": "Событий получено через /api/ingest/events за эту симуляцию"
          },
          "events_accepted": {
            "type": "integer",
            "title": "Events Accepted",
            "description": "Из них новых (остальное — дубли уже загруженных)"
          },
          "critical_sent": {
            "type": "integer",
            "title": "Critical Sent",
            "description": "Уведомлений о новых критических прогнозах разослано в SSE"
          },
          "available_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available From",
            "description": "Первые сутки с прогнозами"
          },
          "available_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available To",
            "description": "Последние сутки с прогнозами"
          },
          "default_day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Day",
            "description": "Сутки по умолчанию: последние полные сутки с прогнозами"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "description": "Что именно делает симуляция"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "active",
          "speed",
          "events_received",
          "events_accepted",
          "critical_sent",
          "note"
        ],
        "title": "SimStateOut"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}